Modified the printed out string

+++ src/misc/List_PrintID.c
@@ -10,7 +10,7 @@
 
     num = List_Num_Entries(l);
 
-    fprintf(stderr,"Set 0x%x:\n",l);
+    fprintf(stderr,"Set 0x%x:\n",(unsigned int) l);
     for (i = 0, k = 0; i < num; i++) {
       entry = List_Entry(l,i);
